// types.ts
export enum Genre {
  FPS = 'First Person Shooter',
  RPG = 'Role Playing Game',
  Racing = 'Racing / Vehicle',
  Simulation = 'Simulation',
  Puzzle = 'Puzzle',
  Platformer = '3D Platformer',
  Arcade = 'Arcade / Action',
  Horror = 'Survival Horror',
  Strategy = 'Strategy / RTS'
}